backend/server/requirements.txt
pyyaml
docker
requests
flask
flask_sqlalchemy
gunicorn
flask_httpauth